update copy, move code, remove image
