+++ static/src/javascripts/projects/common/modules/experiments/tests/contribs-banner-us-eoy/contribs-banner-us-eoy-impeachment-casuals.js
@@ -1,23 +1,20 @@
 // @flow
 import { getSync as geolocationGetSync } from 'lib/geolocation';
 import { acquisitionsBannerUsEoyTemplate } from 'common/modules/commercial/templates/acquisitions-banner-us-eoy';
-import { getDaysLeftInCampaign } from 'common/modules/onward/history';
+import { daysLeftCopy, getDaysLeftBeforeEOY2019 } from './common';
 
-// Code for countdown copy
-const countDownEnd = new Date('Dec 31, 2019 23:59:59').getTime();
-const daysLeft = getDaysLeftInCampaign(countDownEnd);
+const daysLeft = getDaysLeftBeforeEOY2019();
 
 const geolocation = geolocationGetSync();
 const isUS = geolocation === 'US';
 
 const titles = ['At\xa0this\xa0historic\nmoment\xa0for\xa0America'];
-const messageText = `Donald Trump has been impeached – only the third president in history to face this sanction. But the challenges to American democracy do not end today. 2020 will be an epic year – and the need for robust, independent reporting has never been greater. The Guardian relies on your support. Make a year-end gift today from as little as $1. Thank you.`;
+const messageText = `Donald Trump has been impeached – only the third president in history to face this sanction. But the challenges to American democracy do not end here. 2020 will be an epic year – and the need for robust, independent reporting has never been greater. The Guardian relies on your support. Make a year-end gift today from as little as $1. Thank you.`;
 const ctaText = 'Support The Guardian';
 const tickerHeader = 'Help us reach our year-end goal';
-const tickerHeaderVariant = `${daysLeft} days left in 2019`;
 
 export const contributionsBannerUsEoyImpeachmentCasuals: AcquisitionsABTest = {
-    id: 'ContributionsBannerUsEoyImpeachmentCasuals',
+    id: 'ContributionsBannerUsEoyImpeachmentCountDownCasuals',
     campaignId: 'USeoy2019',
     start: '2019-12-16',
     expiry: '2020-1-30',
@@ -31,7 +28,7 @@ export const contributionsBannerUsEoyImpeachmentCasuals: AcquisitionsABTest = {
     idealOutcome: 'NA',
     showForSensitive: true,
     componentType: 'ACQUISITIONS_ENGAGEMENT_BANNER',
-    canRun: () => isUS,
+    canRun: () => daysLeft >= 0 && isUS,
     geolocation,
     variants: [
         {
@@ -56,7 +53,7 @@ export const contributionsBannerUsEoyImpeachmentCasuals: AcquisitionsABTest = {
                 ctaText,
                 template: acquisitionsBannerUsEoyTemplate,
                 hasTicker: true,
-                tickerHeader: tickerHeaderVariant,
+                tickerHeader: daysLeftCopy(daysLeft),
                 bannerModifierClass: 'useoy2019',
             },
         },
